ai_server/requirements.txt
fastapi
uvicorn
tensorflow
Pillow
numpy
pydantic